refactor: test coverage issues
