--- src/pages/welcome/data/data 2.json
@@ -1,43 +1,0 @@
-{
-  "data": [
-    {
-      "image": {
-        "src": "./teamImages/avatar-david.png",
-        "altLocaleIndex": "DavidImage"
-      },
-      "nameLocaleIndex": "DavidName",
-      "quickBioLocaleIndex": "DavidBio",
-      "positionLocaleIndex": "developer",
-      "contacts": {
-        "gitHub": "https://github.com/KuzikevichDavid",
-        "email": "wliorjroo@gmail.com"
-      }
-    },
-    {
-      "image": {
-        "src": "./teamImages/avatar-ilya.png",
-        "altLocaleIndex": "IliyaImage"
-      },
-      "nameLocaleIndex": "IliyaName",
-      "quickBioLocaleIndex": "IliyaBio",
-      "positionLocaleIndex": "developer",
-      "contacts": {
-        "gitHub": "https://github.com/raNYnight",
-        "email": "rany4ka@gmail.com"
-      }
-    },
-    {
-      "image": {
-        "src": "./teamImages/avatar-vlad.png",
-        "altLocaleIndex": "VladImage"
-      },
-      "nameLocaleIndex": "VladName",
-      "quickBioLocaleIndex": "VladBio",
-      "positionLocaleIndex": "developer",
-      "contacts": {
-        "gitHub": "https://github.com/vladklen",
-        "email": "Vlad6097944@gmail.com"
-      }
-    }
-  ]
-}
