--- src/pages/welcome/data/types 2.ts
@@ -1,13 +1,0 @@
-export interface DeveloperData {
-  image: {
-    src: string;
-    altLocaleIndex: string;
-  };
-  nameLocaleIndex: string;
-  positionLocaleIndex: string;
-  quickBioLocaleIndex: string;
-  contacts: {
-    gitHub: string;
-    email: string;
-  };
-}
