+++ src/utils/testUtils/types.ts
@@ -1,7 +1,7 @@
+import { RootState } from '@/store/store';
 import { RenderOptions } from '@testing-library/react';
-import { TestRootState } from './provider';
 
 export interface IExtendedRenderOptions extends RenderOptions {
   initialEntries?: string[];
-  initialState?: TestRootState;
+  initialState?: RootState;
 }
